fix: change pipeline preset default dtype to float16
Update all pipeline presets to use torch.float16 instead of float32:
- supervised_train
- supervised_val
- make_train_pipeline (deprecated)
- make_val_pipeline (deprecated)
- simclr
- byol
- multicrop

float16 is the standard for modern GPU training and matches FFCV's
default behavior.

+++ slipstream/pipelines/byol.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+import torch
+
 from slipstream.decoders.crop import DecodeRandomResizedCrop
 from slipstream.pipelines._common import CROP_OFFSET, _gpu_augmentations, _seed
 from slipstream.transforms import IMAGENET_MEAN, IMAGENET_STD, Normalize, ToTorchImage
@@ -32,7 +34,7 @@ def byol(
     for view_id in range(2):
         stages: list = [
             DecodeRandomResizedCrop(size, seed=_seed(seed, CROP_OFFSET, view_id)),
-            ToTorchImage(device=dev),
+            ToTorchImage(device=dev, dtype=torch.float16),
         ]
 
         if view_id == 0:
